Fixed the app freezing when trying to access light sensor reading while the pump is on

+++ include/light_sensor.h
@@ -4,7 +4,7 @@
 #define _LIGHT_SENSOR_H_
 
 // Convert the voltage reading into lumens (unit for measuring light)
-int Light_getLumens ();
+double Light_getLumens ();
 
 // Set the minimum safety threshold for light level
 void Light_setMinThreshold(int min);
